add paypal service,  aggregator, mock responses

+++ modules/payment/types.go
@@ -3,5 +3,5 @@ package payment
 type Product struct {
 	ID    int64
 	Name  string
-	Price int64
+	Price int
 }
